requirements.txt
torch
torchvision
torchaudio
openai-whisper
flask
werkzeug